requirements.txt
fastapi
uvicorn
openai-whisper
pydub
python-multipart
requests
openai
python-dotenv
# Audio Enhancement Libraries
noisereduce>=3.0.0
librosa>=0.10.0
scipy>=1.10.0
webrtcvad>=2.0.10
numpy>=1.24.0